SPの追加、削除、全件取得の実装
